Fixed attachmenta preview. Adopted for iOS 13. Cleaned up the code.
